GFF3 feature exporter
$  export_gff3_feature.pl  /path/to/some.gff3  gene  ID  YAR003W

There are 4 arguments here that correspond to values in the GFF3
columns.  In this case, your script should read the path to a GFF3
file, find any gene (column 3) which has an ID=YAR003W (column 9).
When it finds this, it should use the coordinates for that feature
(columns 4, 5 and 7) and the FASTA sequence at the end of the document
to return its FASTA sequence.